Solving the problem 10 in python

+++ Problem_7/main.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-bool isPrime(int number)
+bool isPrime(size_t number)
 {
     if (number == 2)
         return true;
@@ -8,7 +8,7 @@ bool isPrime(int number)
     if  (number % 2 == 0)
         return false;
 
-    for (int i = 3; i <= number / 3; i+=2)
+    for (size_t i = 3; i <= number / 3; i+=2)
     {
         if (number % i == 0)
             return false;
@@ -19,12 +19,12 @@ bool isPrime(int number)
 
 int main(int argc, char const *argv[])
 {
-    int amount = 10001;
+    size_t amount = 10001;
 
     // taking into account that the first prime number is 2 and
     // the other are odd.
-    int count = 1;
-    int i = 3;
+    size_t count = 1;
+    size_t i = 3;
     while (count < amount)
     {
         if (isPrime(i))
